Write else-related labels regarldess if else-block exists
The reason for this is because we have a goto for the if-block (i.e. if the expression is false, go to the else branch). Previously, if we did not have an else-branch, the label would not have existed and there would be a compiler exception.

I've decided to write the else-related instructions anyway, it's not optimal because there are now extra instructions we don't always need, but that's the problem for another time.